Implementar testes e fazer alterações

+++ PO-008/EmpresaDeEnergia/src/entidades/Pagamento.java
@@ -4,10 +4,10 @@ import java.util.Calendar;
 
 public class Pagamento {
     private int idFatura;
-    private float valor;
+    private double valor;
     private Calendar data;
 
-    public Pagamento(int idFatura, float valor, Calendar data) {
+    public Pagamento(int idFatura, double valor, Calendar data) {
         this.idFatura = idFatura;
         this.valor = valor;
         this.data = data;
@@ -17,7 +17,7 @@ public class Pagamento {
         return idFatura;
     }
 
-    public float getValor() {
+    public double getValor() {
         return valor;
     }
